work

+++ src/main/java/com/pruebatec/application/repositories/VoteCountRepository.java
@@ -6,10 +6,9 @@ import java.util.List;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.CrossOrigin;
 
 @Repository
-public interface VoteCountRepository extends CrudRepository<VoteCount, Long> {
+public interface VoteCountRepository extends CrudRepository<VoteCount, Integer> {
     @Query("SELECT v.musicStyle, COUNT(v) FROM VoteCount v GROUP BY v.musicStyle")
     List<Object[]> countVotesByMusicStyle();
 }
